agent/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
requests==2.31.0
python-socketio==5.9.0
sounddevice>=0.4.7
numpy>=1.24
boto3>=1.28.0
pytest>=7.0.0
keyboard>=0.13.5
obs-websocket-py>=0.6.5
